requirements.txt
autopep8==1.6.0
climage==0.1.3
commonmark==0.9.1
culour==0.2
kdtree==0.16
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
pycodestyle==2.8.0
Pygments==2.12.0
rich==12.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
toml==0.10.2
